refactor(java): remove unused methods and imports WorkLogConfig new class WorkLogDisplay
- Remove unused date and mapping imports (e.g., `Set`, `TreeMap`, `IsoFields`, `WeekFields`, `Locale`).
- Delete `endOfMonth`, `nzHolidays2025`, and unused month constants.
- Remove redundant console logging methods: `printDataStructures()`, `printHolidays()`, and `printDaysUntilEndofMonth()`.
- Clean up the `run()` execution block by removing the calls to the deleted printing methods.
- Add an end-of-class comment for better structural readability.
